requirements.txt
Django==1.10.4
Markdown==2.6.7
MySQL-python==1.2.5
argparse==1.2.1
django-filter==1.0.1
djangorestframework==3.5.3
pony==0.7
wsgiref==0.1.2
